data-platform/share-layer/bucket/requirements.txt
# Database connectivity
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# MinIO client
minio==7.2.3

# Data processing
pandas==2.1.4
numpy==1.26.2

# JSON handling
orjson==3.9.15

# Environment management
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Async support
asyncio==3.4.3
aiofiles==23.2.1

# Progress bars
tqdm==4.66.3

# Date handling
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Type checking
mypy==1.7.1
types-python-dateutil==2.8.19.14